wasip1: make pure Go resolver the default
